+++ src/pages/implementation-references/react-hook-form.tsx
@@ -3,9 +3,12 @@ import { useEffect } from 'react';
 import { SubmitHandler, useForm, FormProvider } from 'react-hook-form';
 import { TextInput } from '@/components/forms/text-input';
 import { Button } from 'antd';
+import { FileDropzone } from '@/components/forms/file-dropzone';
+import { FileWithPath } from 'file-selector';
 interface ISampleForm {
   usuario: string;
   projeto: string;
+  file: FileWithPath[];
 }
 
 const ReactHookFormPage: NextPage = () => {
@@ -26,6 +29,7 @@ const ReactHookFormPage: NextPage = () => {
         <TextInput name="usuario" placeholder="usuario" />
         <TextInput name="projeto" placeholder="projeto" />
         <h1>{formChanges.value}</h1>
+        <FileDropzone name="file" />
         <Button type="primary" htmlType="submit">
           Enviar
         </Button>